Move mocked_files to unit_tests/conftest.py
