+++ tests/ert/unit_tests/config/test_rft_config.py
@@ -1,7 +1,5 @@
-import io
-import os
 from datetime import datetime
-from io import BytesIO, StringIO
+from io import BytesIO
 from pathlib import Path
 from typing import Any, cast
 
@@ -23,10 +21,6 @@ from ert.config.parsing import ConfigValidationError, ObservationType
 from ert.config.rft_config import _get_zonemap, _read_egrid
 from tests.ert.rft_generator import cell_start, float_arr
 
-original_open = open
-original_io_open = io.open
-original_stat = os.stat
-
 
 @pytest.fixture(autouse=True)
 def _clear_rft_caches():
@@ -37,42 +31,6 @@ def _clear_rft_caches():
     yield
     _read_egrid.cache_clear()
     _get_zonemap.cache_clear()
-
-
-@pytest.fixture
-def mocked_files(mocker):
-    mocked_files = {}
-
-    def _fresh_buffer(data):
-        if isinstance(data, bytes):
-            return BytesIO(data)
-        return StringIO(data)
-
-    def mock_open(*args, **kwargs):
-        path = args[0] if args else kwargs.get("file")
-        data = mocked_files.get(str(path))
-        if data is not None:
-            return _fresh_buffer(data)
-        return original_open(*args, **kwargs)
-
-    def mock_io_open(*args, **kwargs):
-        path = args[0] if args else kwargs.get("file")
-        data = mocked_files.get(str(path))
-        if data is not None:
-            return _fresh_buffer(data)
-        return original_io_open(*args, **kwargs)
-
-    def mock_stat(*args, **kwargs):
-        path = args[0] if args else kwargs.get("path")
-        if str(path) in mocked_files:
-            return os.stat_result([0x777, *([1] * 10)])
-        return original_stat(*args, **kwargs)
-
-    mocker.patch("builtins.open", mock_open)
-    mocker.patch("io.open", mock_io_open)
-    mocker.patch("os.stat", mock_stat)
-
-    return mocked_files
 
 
 @pytest.fixture
